refactor(data): refactored/moved dataclasses to appropriate modules
